requirements.txt
anyio>=3.6.2
appnope>=0.1.3
argon2-cffi>=21.3.0
argon2-cffi-bindings>=21.2.0
asttokens>=2.1.0
attrs>=22.1.0
backcall>=0.2.0
backports.functools-lru-cache>=1.6.4
beautifulsoup4>=4.11.1
bleach>=5.0.1
cachetools>=5.3.0
certifi>=2022.9.24
cffi>=1.15.1
charset-normalizer>=3.0.1
click>=8.1.3
contourpy>=1.0.6
cycler>=0.11.0
debugpy>=1.6.3
decorator>=5.1.1
defusedxml>=0.7.1
entrypoints>=0.4
executing>=1.2.0
fastjsonschema>=2.16.2
flit_core>=3.7.1
fonttools>=4.38.0
google-api-core>=2.11.0
google-api-python-client>=2.74.0
google-auth>=2.16.0
google-auth-httplib2>=0.1.0
googleapis-common-protos>=1.58.0
greenlet>=2.0.0.post0
httplib2>=0.21.0
idna>=3.4
importlib-metadata>=5.0.0
importlib-resources>=5.10.0
ipykernel>=6.17.0
ipython>=8.11.0
ipython-genutils>=0.2.0
ipywidgets>=8.0.2
jedi>=0.18.1
Jinja2>=3.1.2
jsonschema>=4.17.0
jupyter_client>=7.4.4
jupyter-console>=6.4.4
jupyter_core>=4.11.2
jupyter-server>=1.21.0
jupyterlab-pygments>=0.2.2
jupyterlab-widgets>=3.0.3
kiwisolver>=1.4.4
MarkupSafe>=2.1.1
mistune>=2.0.4
msgpack>=1.0.4
munkres>=1.1.4
mypy>=0.990
mypy-extensions>=0.4.3
nbclassic>=0.4.8
nbclient>=0.7.0
nbconvert>=7.2.3
nbformat>=5.7.0
nest-asyncio>=1.5.6
notebook>=6.5.2
notebook_shim>=0.2.2
packaging>=21.3
pandocfilters>=1.5.0
parso>=0.8.3
pexpect>=4.8.0
pickleshare>=0.7.5
Pillow>=9.2.0
pip>=23.0.1
pkgutil_resolve_name>=1.3.10
ply>=3.11
prometheus-client>=0.15.0
prompt-toolkit>=3.0.31
protobuf>=4.21.12
psutil>=5.9.3
ptyprocess>=0.7.0
pure-eval>=0.2.2
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
pycparser>=2.21
Pygments>=2.13.0
pynvim>=0.4.3
pyparsing>=3.0.9
PyQt5-sip>=12.11.0
pyrsistent>=0.19.2
python-dateutil>=2.8.2
pytz>=2022.6
PyYAML>=6.0
pyzmq>=24.0.1
qtconsole>=5.4.0
QtPy>=2.2.1
requests>=2.28.2
rsa>=4.9
Send2Trash>=1.8.0
setuptools>=67.4.0
sip>=6.7.3
six>=1.16.0
sniffio>=1.3.0
soupsieve>=2.3.2.post1
stack-data>=0.6.0
terminado>=0.17.0
tinycss2>=1.2.1
toml>=0.10.2
tomli>=2.0.1
tornado>=6.2
traitlets>=5.5.0
types-PyYAML>=6.0.12.8
typing_extensions>=4.4.0
uritemplate>=4.1.1
urllib3>=1.26.14
wcwidth>=0.2.5
webencodings>=0.5.1
websocket-client>=1.4.1
wheel>=0.38.4
widgetsnbextension>=4.0.3
zipp>=3.10.0
